+++ tenpy/linalg/np_conserved.py
@@ -203,22 +203,25 @@ class Array:
             raise ValueError("setstate with incompatible type of state")
 
     @classmethod
-    def from_ndarray_trivial(cls, data_flat, dtype=np.float64):
+    def from_ndarray_trivial(cls, data_flat, dtype=None):
         """convert a flat numpy ndarray to an Array with trivial charge conservation.
 
         Parameters
         ----------
         data_flat : array_like
             The data to be converted to a Array.
-        dtype : type | string
-            The data type of the array entries. Defaults to ``np.float64``.
+        dtype : ``np.dtype``
+            The data type of the array entries. Defaults to dtype of `data_flat`.
 
         Returns
         -------
         res : :class:`Array`
             An Array with data of data_flat.
         """
-        data_flat = np.array(data_flat, dtype)
+        data_flat = np.asarray(data_flat)  # unspecified dtype
+        if dtype is None:
+            dtype = data_flat.dtype
+        data_flat = data_flat.astype(dtype, copy=False)
         chinfo = ChargeInfo()
         legs = [LegCharge.from_trivial(s, chinfo) for s in data_flat.shape]
         res = cls(legs, dtype)
@@ -239,8 +242,8 @@ class Array:
             The shape has to be compatible with legcharges.
         legcharges : list of :class:`LegCharge`
             The leg charges for each of the legs. The :class:`ChargeInfo` is read out from it.
-        dtype : ``np.dtype`` | string
-            The data type of the array entries. Defaults to dtype of data_flat.
+        dtype : ``np.dtype``
+            The data type of the array entries. Defaults to dtype of `data_flat`.
         qtotal : None | charges
             The total charge of the new array.
         cutoff : float
@@ -261,8 +264,8 @@ class Array:
         data_flat = np.asarray(data_flat)  # unspecified dtype
         if dtype is None:
             dtype = data_flat.dtype
-        res = cls(legcharges, dtype, qtotal)  # without any data
         data_flat = data_flat.astype(dtype, copy=False)
+        res = cls(legcharges, dtype, qtotal)  # without any data
         if res.shape != data_flat.shape:
             raise ValueError("Incompatible shapes: legcharges {0!s} vs flat {1!s} ".format(
                 res.shape, data_flat.shape))
@@ -3850,8 +3853,9 @@ def _inner_worker(a, b, do_conj):
     """Full contraction of `a` and `b` with axes in matching order."""
     calc_dtype, res_dtype = _find_calc_dtype(a.dtype, b.dtype)
     res = res_dtype.type(0)
-    if any(a.chinfo.make_valid(a.qtotal + b.qtotal) != 0):
-        return res  # can't have blocks to be contracted
+    check_qtotal = b.qtotal - a.qtotal if do_conj else b.qtotal + a.qtotal
+    if np.any(a.chinfo.make_valid(check_qtotal) != 0):
+        return res # can't have blocks to be contracted.
     if a.stored_blocks == 0 or b.stored_blocks == 0:
         return res  # also trivial
     a = a.astype(calc_dtype, False)
@@ -3878,6 +3882,7 @@ def _inner_worker(a, b, do_conj):
     for i, j in _iter_common_sorted(a_qdata, b_qdata):
         res += blas_dot(a_data[i], b_data[j])
         # same as res += np.inner(a_data[i].reshape((-1, )), b_data[j].reshape((-1, )))
+        # (or with complex conj if 'do_conj')
     return res